ignore anything in the cookie which doesn't look like key=value
Previously, this would cause empty key fields to be set to whatever the
string was. So a cookie of `key=value;Foo` would make an object with and
empty key and value of `Foo`. While this is not a valid cookie, we don't
want to create these sort of partial objects.

close #13