Update index.js, #serialize so it accepts opt.expires as a date string or a unix number
Recently removed the cookie parser middleware for express' own. the format is similar, the only difference is that `expires` it is not returned as a Date Object, instead it is a date string. 

Wrapping `opt.expires` with `new Date`, guarantees that calling the toUTCString method does not throw. This eases the transition from the other cookie middleware to this one.

TypeError: undefined is not a function
    at Object.serialize (/home/ascari/node_modules/express-session/node_modules/cookie/index.js:25:58)
    at setcookie (/home/ascari/node_modules/express-session/index.js:577:21)
    at ServerResponse.<anonymous> (/home/ascari/node_modules/express-session/index.js:200:7)
    at ServerResponse.writeHead (/home/ascari/node_modules/express-session/node_modules/on-headers/index.js:44:16)
    at ServerResponse._implicitHeader (_http_server.js:172:8)
    at ServerResponse.OutgoingMessage.write (_http_outgoing.js:423:10)
    at writetop (/home/ascari/node_modules/express-session/index.js:248:26)
    ..etc