fix(serialize): added regex validation for cookie name
A vulnerability exists because the cookie name was not validated nor encoded to avoid 'separator' characters which lead to the ability for cookie names to be exploited for arbitrary cookie injection by passing in delimiters in the cookie name string.

This fix adds a regex test based on [RFC 6265 sec 4.1.1][1] and [RFC 2616 sec 2.2][2] to check for valid characters allowed in cookie names. It throws an error if the criteria are not met as opposed to sanitizing the value which could lead to silent bugs and the potential for cookie name collision as different name strings could be santized to the same resultant value.

Tests also provided to ensure valid cookie names work, invalide ones error as well as a couple attack strings.

[1]: https://datatracker.ietf.org/doc/html/rfc2616#section-2.2
[2]: https://datatracker.ietf.org/doc/html/rfc2616#section-2.2